QS
- fixed some update bugs
- pr commentaries
