+++ frontend/src/database/types/internal/DbEntryMethaData.tsx
@@ -1,4 +1,5 @@
 export interface DbEntryMethaData {
+  ok: Boolean;
   id: string;
   rev: string;
 }
