+++ frontend/src/database/types/public/Comment.tsx
@@ -1,7 +1,7 @@
 export interface Comment {
   _id?: string;
   author: string;
-  date: string;
+  date: string; //ISO-8601
   content: string;
   likes: number;
   dislikes: number;
