+++ frontend/src/database/types/public/Post.tsx
@@ -3,7 +3,7 @@ import { Comment } from './Comment';
 export interface Post {
   _id?: string;
   author: string;
-  date: string;
+  date: string; //ISO-8601
   content: string;
   likes: Number;
   dislikes: Number;
